pyproject.toml
[tool.poetry]
name = "dltoolkit"
version = "0.1.0"
description = "A useful toolkit for Deep Learning and Python in general"
authors = ["Kharinaev Artyom <akharinaev@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Kharinaev/dltoolkit"

[tool.poetry.dependencies]
python = "3.10.13"

PyYAML = "^6.0"
toml = "^0.10.2"

numpy = "^1.26.0"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
scikit-learn = "^1.3.1"
torch = "2.0.1"
torchvision = "0.15.2"
Pillow = "9.4.0"
tqdm = "^4.65.0"
onnx = "1.15.0"
onnxruntime = "1.16.3"
tritonclient = "2.41.0"
gevent = "23.9.1"
geventhttpclient = "2.0.11"

[tool.poetry.dev-dependencies]
pre-commit = "^3.4.0"
dvc = "3.30.3"
dvc-gdrive = "2.20.0"
fire = "0.5.0"
hydra-core = "1.3.2"
mlflow = "2.9.1"
catalyst = "22.4"
albumentations = "1.3.1"
gitpython = "3.1.40"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["dltoolkit"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
